move actors to exports_test.go
